// src/_proto/Protos/users.ts
export interface UserSettings {
    /**
     * @generated from protobuf field: bool autoUnlockChapters = 1;
     */
    autoUnlockChapters: boolean;
    /**
     * @generated from protobuf field: bool is2faEnabled = 2 [json_name = "is2faEnabled"];
     */
    is2FaEnabled: boolean;
    /**
     * @generated from protobuf field: bool allowMarketing = 3;
     */
    allowMarketing: boolean;
}
/**
 * @generated from protobuf message wuxiaworld.api.v2.UserPermissions
 */
export interface UserPermissions {
    /**
     * @generated from protobuf field: repeated wuxiaworld.api.v2.Role roles = 1;
     */
    roles: Role[];
    /**
     * @generated from protobuf field: repeated wuxiaworld.api.v2.ObjectItem objects = 2;
     */
    objects: ObjectItem[];
}
/**
 * @generated from protobuf message wuxiaworld.api.v2.UserItem
 */
export interface UserItem {
    /**
     * @generated from protobuf field: string id = 1;
     */
    id: string;
    /**
     * @generated from protobuf field: string userName = 2;
     */
    userName: string;
    /**
     * @generated from protobuf field: google.protobuf.StringValue avatarUrl = 3;
     */
    avatarUrl?: StringValue;
    /**
     * @generated from protobuf field: google.protobuf.StringValue email = 4;
     */
    email?: StringValue;
    /**
     * @generated from protobuf field: google.protobuf.Timestamp joinDate = 5;
     */
    joinDate?: Timestamp;
    /**
     * @generated from protobuf field: bool isVipActive = 6;
     */
    isVipActive: boolean;
    /**
     * @generated from protobuf field: wuxiaworld.api.v2.VipItem vip = 7;
     */
    vip?: VipItem;
    /**
     * @generated from protobuf field: google.protobuf.StringValue flair = 8;
     */
    flair?: StringValue;
    /**
     * @generated from protobuf field: wuxiaworld.api.v2.UserSettings settings = 9;
     */
    settings?: UserSettings;
    /**
     * @generated from protobuf field: wuxiaworld.api.v2.UserPermissions permissions = 10;
     */
    permissions?: UserPermissions;
    /**
     * @generated from protobuf field: bool isLegacy = 11;
     */
    isLegacy: boolean;
    /**
     * @generated from protobuf field: bool isLegacyVip = 12;
     */
    isLegacyVip: boolean;
}
// ...
class UserSettings$Type extends MessageType<UserSettings> {
    constructor() {
        super("wuxiaworld.api.v2.UserSettings", [
            { no: 1, name: "autoUnlockChapters", kind: "scalar", T: 8 /*ScalarType.BOOL*/ },
            { no: 2, name: "is2faEnabled", kind: "scalar", jsonName: "is2faEnabled", T: 8 /*ScalarType.BOOL*/ },
            { no: 3, name: "allowMarketing", kind: "scalar", T: 8 /*ScalarType.BOOL*/ }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message wuxiaworld.api.v2.UserSettings
 */
export const UserSettings = new UserSettings$Type();
// @generated message type with reflection information, may provide speed optimized methods
class UserPermissions$Type extends MessageType<UserPermissions> {
    constructor() {
        super("wuxiaworld.api.v2.UserPermissions", [
            { no: 1, name: "roles", kind: "message", repeat: 1 /*RepeatType.PACKED*/, T: () => Role },
            { no: 2, name: "objects", kind: "message", repeat: 1 /*RepeatType.PACKED*/, T: () => ObjectItem }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message wuxiaworld.api.v2.UserPermissions
 */
export const UserPermissions = new UserPermissions$Type();
// @generated message type with reflection information, may provide speed optimized methods
class UserItem$Type extends MessageType<UserItem> {
    constructor() {
        super("wuxiaworld.api.v2.UserItem", [
            { no: 1, name: "id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "userName", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "avatarUrl", kind: "message", T: () => StringValue },
            { no: 4, name: "email", kind: "message", T: () => StringValue },
            { no: 5, name: "joinDate", kind: "message", T: () => Timestamp },
            { no: 6, name: "isVipActive", kind: "scalar", T: 8 /*ScalarType.BOOL*/ },
            { no: 7, name: "vip", kind: "message", T: () => VipItem },
            { no: 8, name: "flair", kind: "message", T: () => StringValue },
            { no: 9, name: "settings", kind: "message", T: () => UserSettings },
            { no: 10, name: "permissions", kind: "message", T: () => UserPermissions },
            { no: 11, name: "isLegacy", kind: "scalar", T: 8 /*ScalarType.BOOL*/ },
            { no: 12, name: "isLegacyVip", kind: "scalar", T: 8 /*ScalarType.BOOL*/ }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message wuxiaworld.api.v2.UserItem
 */
export const UserItem = new UserItem$Type();